requirements.txt
crewai==0.1.0
langchain==0.1.0
openai==1.3.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
selenium==4.15.0
amadeus==9.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
apscheduler==3.10.4
aiohttp==3.9.0
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
plotly==5.18.0
matplotlib==3.8.0
redis==5.0.1
stripe==7.8.0
google-api-python-client==2.111.0
icalendar==5.0.11
yfinance==0.2.33
textblob==0.17.1
pytz==2023.3
Pillow==10.1.0